ARM64: Add named branch target support in SymbolTable
This patch enhances branch diassembly support in ARM64.

This patch first purifies ARM64 branch disassembly so that the decoded
instructions can be re-compiled by GNU assembler.
 bl #address to bl symbol
 b.ne <symbol+offset> to b.ne #offset	// <symbol+offset>

This patch then further enhances "#offset" to "target" label (in the format
of symbol_offset). The idea of achieving this is to mark the branch target
in the symbol table during early disassembling. This feature is only
enabled for BaseSimpleCPU when TRACING_ON is enabled as we may only use it
to dump checkpoint assemblies.

However there is no user of markTarget() introduced in this patch. The
feature will be used by further commits.

Change-Id: If223e52ae3782af161adee133857ada13d796739
Signed-off-by: Lv Zheng <zhenglv@hotmail.com>