ARM64: Fix register width in disassembly for Load/Store (ll/sc and cas) disassembly
For some Load/Store instructions including:
 - Load-Exclusive/Store-Exclusive instructions
 - Non-exclusive Load-Acquire and Store-Release instructions
 - Exclusive Load-Acquire and Store-Release instructions
the width of Rd could be 32 or 64 bits, depending on the size field of
the machine code, the width of Rs is always 32 bits, and the width of Rn
(base register) is always 64 bits.
The origin GEM5 does not give the correct disassembly. This patch fixes
the problem.

Change-Id: I3b38a8637b84af0f4e87df50226eaa72c59b1dd6
Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>
Signed-off-by: Lv Zheng <zhenglv@hotmail.com>