ARM64: Fix register width of disassembly for Load/Store (exclusive or not)
For some Load/Store instructions including:
- Load-Exclusive/Store-Exclusive instructions
- Non-exclusive Load-Acquire and Store-Release instructions
- Exclusive Load-Acquire and Store-Release instructions
the width of Rd could be 32 or 64 bits, depending on bit31:30 of the machine
code, the width of Rs is always 32 bits, and the width of Rn (base register)
is always 64 bits. The origin GEM5 does not give the correct disassembly.
This patch fixes the problem.

Change-Id: If4f9b533c02b669d5e447be0d198b6fd167a53eb
Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>