arch-riscv: Fix immediate decoding for integer shift immediate instructions
The "shamt" in integer shift immediate instructions is an unsigned immediate
encoded in bits[25:20]. While the original Gem5 uses bits[31:20] as an int64_t.
This patch fixes the problem by adding for these instructions a new format
IShiftOp in which the immediate type is changed to uint64_t and the immediate
value is assigned with the correct bitfields SHAMT6.

The instructions affected include:
- Shift Left Logical Immediate, slli
- Shift Right Logical Immediate, srli
- Shift Right Arithmetic Immediate, srai
- Shift Left Logical Word Immediate, slliw
- Shift Right Logical Word Immediate, srliw
- Shift Right Arithmetic Word Immediate, sraiw

Change-Id: Iad34ccd036c11630409f84f6de2b939224e100e6
Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>