arch-riscv: Fix disassembling for atomic instructions
The original Gem5 does not give correct disassembly for atomic instructions,
which are implemented with one or two micro instructions. The correct register
indices are not decoded until subsequent micro instruction is processed.
This patch fixes the problem by decoding the register indices and other
properties (aq and rl) from the machine code directly in the disassembling
function.

Change-Id: I2cdaf0b3c48ff266f19ca707a5de48c9050b3897
Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>