arch-riscv: Add named branch target support in SymbolTable
This patch enhances branch diassembly support in RISCV.

TODO:
This patch first purifies RISCV branch disassembly so that the decoded
instructions can be re-compiled by GNU assembler.

TODO:
This patch then further enhances "#offset" to "target" label (in the format
of symbol_offset). The idea of achieving this is to mark the branch target
in the symbol table during early disassembling. This feature is only
enabled for BaseSimpleCPU when TRACING_ON is enabled as we may only use it
to dump checkpoint assemblies.

However there is no user of markTarget() introduced in this patch. The
feature will be used by further commits.

Change-Id: I10706bc69829672d55bef43d68ac243a4f62c845
Signed-off-by: Lv Zheng <zhenglv@hotmail.com>
Signed-off-by: Ian Jiang <ianjiang.ict@gmail.com>